Add integration test for rich messages when Comments app is disabled
The resolution of display names in rich messages is affected by whether
the Comments app is enabled or not, so this is now tested explicitly by
an integration test.

Once the Comments app is disabled in a test it is not automatically
enabled again before the next tests are run, so it is now explicitly
enabled before every scenario in the rich message feature (even if
currently it is not strictly needed, as the test that disables it is the
last one whose result depends on the Comments app being enabled or not,
but it was done anyway for correctness).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>